EOS-25300 Resolve all issues for multi-client run

+++ performance/PerfPro/roles/benchmark/files/PerfProBenchmark/system_monitoring/stopSystemMonitoring.py
@@ -1,3 +1,6 @@
 #!/usr/bin/env python3
 import os
-os.remove("pidfile")
+if os.path.isfile("pidfile"):
+    os.remove("pidfile")
+else:
+    print("No systemstats are running! Nothing to stop. Exiting")
